django/requirements.txt
Django==1.11.6
Pillow==4.3.0
mysqlclient==1.3.12
django-import-export==0.5.1
uwsgi==2.0.15
django-allauth==0.33.0
django-phonenumber-field==1.3.0
django-countries==5.0
django_money==0.12
